16-may-almost-done

+++ src/models/ReceiptTransaction.model.js
@@ -18,22 +18,22 @@ module.exports = (sequelize, Sequelize) => {
         type: Sequelize.DATE,
       },
       amount: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
       },
       mora: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
       },
       discount: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
       },
       discount_interest: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
       },
       discount_mora: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
       },
       total_paid: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DOUBLE,
       },
       cashback: {
         type: Sequelize.DOUBLE,
